Refactor MCP server into package structure
