+++ anki_mcp/actions/__init__.py
@@ -1,0 +1,5 @@
+"""Пакет с MCP-действиями."""
+
+from .search import search
+
+__all__ = ["search"]
