+++ anki_mcp/services/__init__.py
@@ -1,0 +1,6 @@
+"""Прикладные сервисы MCP-сервера."""
+
+from . import anki, search
+
+
+__all__ = ["anki", "search"]
